+++ src/SpatialUnitTests/Euclidean/UnitVector3DTests.cs
@@ -1,6 +1,7 @@
 ﻿namespace MathNet.Spatial.UnitTests.Euclidean
 {
     using System;
+    using System.Globalization;
     using System.IO;
     using System.Xml;
     using System.Xml.Serialization;
@@ -31,6 +32,62 @@
             Assert.Throws<ArgumentException>(() => new UnitVector3D(new[] { 1.0, 2, 3, 4 }));
         }
 
+        [TestCase("1,0; 0; 0,0", 1, 0, 0)]
+        [TestCase("0; 1,0; 0,0", 0, 1, 0)]
+        [TestCase("0; 0,0; 1,0", 0, 0, 1)]
+        [TestCase("1.0; 0; 0.0", 1, 0, 0)]
+        [TestCase("0; 1.0; 0.0", 0, 1, 0)]
+        [TestCase("0; 0.0; 1.0", 0, 0, 1)]
+        public void Parse(string text, double expectedX, double expectedY, double expectedZ)
+        {
+            Assert.AreEqual(true, UnitVector3D.TryParse(text, out var p));
+            Assert.AreEqual(expectedX, p.X);
+            Assert.AreEqual(expectedY, p.Y);
+            Assert.AreEqual(expectedZ, p.Z);
+
+            p = UnitVector3D.Parse(text);
+            Assert.AreEqual(expectedX, p.X);
+            Assert.AreEqual(expectedY, p.Y);
+            Assert.AreEqual(expectedZ, p.Z);
+
+            p = UnitVector3D.Parse(p.ToString());
+            Assert.AreEqual(expectedX, p.X);
+            Assert.AreEqual(expectedY, p.Y);
+            Assert.AreEqual(expectedZ, p.Z);
+        }
+
+        [TestCase("1,0; 0; 0,0", 1, 0, 0)]
+        [TestCase("0; 1,0; 0,0", 0, 1, 0)]
+        [TestCase("0; 0,0; 1,0", 0, 0, 1)]
+        public void ParseSwedish(string text, double expectedX, double expectedY, double expectedZ)
+        {
+            var culture = CultureInfo.GetCultureInfo("sv");
+            Assert.AreEqual(true, UnitVector3D.TryParse(text, culture, out var p));
+            Assert.AreEqual(expectedX, p.X);
+            Assert.AreEqual(expectedY, p.Y);
+            Assert.AreEqual(expectedZ, p.Z);
+
+            p = UnitVector3D.Parse(text, culture);
+            Assert.AreEqual(expectedX, p.X);
+            Assert.AreEqual(expectedY, p.Y);
+            Assert.AreEqual(expectedZ, p.Z);
+
+            p = UnitVector3D.Parse(p.ToString(culture));
+            Assert.AreEqual(expectedX, p.X);
+            Assert.AreEqual(expectedY, p.Y);
+            Assert.AreEqual(expectedZ, p.Z);
+        }
+
+        [TestCase("1; 2; 3")]
+        [TestCase("1.2")]
+        [TestCase("1,2; 2.3; 3")]
+        [TestCase("1; 2; 3; 4")]
+        public void ParseFails(string text)
+        {
+            Assert.AreEqual(false, UnitVector3D.TryParse(text, out _));
+            Assert.Throws<FormatException>(() => UnitVector3D.Parse(text));
+        }
+
         [Test]
         public void ToDenseVector()
         {
@@ -44,8 +101,12 @@
             Assert.AreEqual(3 / l, denseVector[2], 1e-6);
         }
 
-        [TestCase("1, 2, 3", "1, 2, 3", 1e-4, true)]
-        [TestCase("1, 2, 3", "4, 5, 6", 1e-4, false)]
+        [TestCase("1, 0, 0", "1, 0, 0", 1e-4, true)]
+        [TestCase("0, 1, 0", "0, 1, 0", 1e-4, true)]
+        [TestCase("0, 0, 1", "0, 0, 1", 1e-4, true)]
+        [TestCase("1, 0, 0", "0, 1, 0", 1e-4, false)]
+        [TestCase("0, 1, 0", "1, 0, 0", 1e-4, false)]
+        [TestCase("0, 0, 1", "0, 1, 0", 1e-4, false)]
         public void Equals(string p1s, string p2s, double tol, bool expected)
         {
             var v1 = UnitVector3D.Parse(p1s);